requirements.txt
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.0
ccxt==4.1.86
PyYAML==6.0.1
pandas==2.2.2
